+++ admin/functions-admin.php
@@ -20,8 +20,12 @@ function arve_action_admin_init_setup_messages() {
 
 		$pro_ad_message .= file_get_contents( ARVE_PATH . 'admin/pro-ad.html' );
 
-		new ARVE_Admin_Notice_Factory( 'arve_dismiss_pro_notice', $pro_ad_message, true );
+		new ARVE_Admin_Notice_Factory( 'arve_dismiss_pro_notice', $pro_ad_message );
 	}
+
+	$survery_message = __( '<p>Hi, this is Nico(las Jonas) the author of the ARVE plugin. I am in a delemma because there was offered a advertising deal that is hard to resist but I am commited to this plugin and to not want to damage its reputation and annoy you. If you have a few minutes please <a href="https://nextgenthemes.com/advertising-offer/" target="_blank">share your opinion</a>.</p>', ARVE_SLUG );
+
+	#new ARVE_Admin_Notice_Factory( 'arve_dismiss_pro_notice', $survery_message );
 }
 
 function arve_add_tinymce_plugin( $plugin_array ) {
@@ -163,7 +167,7 @@ function arve_add_media_button() {
 
 	printf(
 		"<div id='arve-thickbox' style='display:none;'><p>$p1</p><p>$p2</p><p>$p3</p></div>",
-		nextgenthemes_admin_install_search_url( 'Shortcode+UI' ),
+		\nextgenthemes\admin\plugin_install_search_url( 'Shortcode+UI' ),
 		esc_url( 'https://nextgenthemes.com/plugins/advanced-responsive-video-embedder-pro/documentation/' )
 	);